+++ src/math/float.rs
@@ -7,7 +7,10 @@ use std::{
     str::FromStr,
 };
 
-use crate::private::Sealed;
+use crate::{
+    math::cast::{CastAsNumeric, CastFromNumeric},
+    private::Sealed,
+};
 
 /// Takes square root using the Babylonian Method using 40 iterations.
 // Relevant discussion: <https://www.codeproject.com/Articles/69941/Best-Square-Root-Method-Algorithm-Function-Precisi>
@@ -31,8 +34,10 @@ pub(crate) const fn sqrt_baby(n: f64) -> f64 {
 
 /// Trait for providing generic functionality over floating point numbers.
 pub trait Float:
-    Sub<Output = Self>
+    CastAsNumeric
+    + CastFromNumeric
     + Add<Output = Self>
+    + Sub<Output = Self>
     + Mul<Output = Self>
     + Div<Output = Self>
     + Neg<Output = Self>
@@ -85,9 +90,6 @@ pub trait Float:
     /// Generic exponential for [`Float`]
     #[must_use]
     fn exp(self) -> Self;
-
-    /// Use a primitive cast to convert a usize to the [`Float`]
-    fn usize_as_self(a: usize) -> Self;
 }
 
 /// Implement [`Float`] for multiple floating point primitive types
@@ -135,12 +137,6 @@ macro_rules! impl_float {
             #[inline]
             fn exp(self) -> Self {
                 self.exp()
-            }
-
-            #[inline]
-            #[allow(clippy::cast_precision_loss)]
-            fn usize_as_self(a: usize) -> $ty {
-                a as $ty
             }
         } )*